+++ apps/api/src/entities/getEntityEmbeds.ts
@@ -1,12 +1,14 @@
 import { isLink, Links } from "@phylopic/api-models"
+import { S3Client } from "@aws-sdk/client-s3"
 import { isString } from "@phylopic/utils"
-import { ClientBase } from "pg"
 import APIError from "../errors/APIError"
-import selectEntitiesJSONFromLinks from "./selectEntitiesJSONFromLinks"
-import selectEntityJSONFromHRef from "./selectEntityJSONFromHRef"
+import getEntitiesJSONFromLinks from "./getEntitiesJSONFromLinks"
+import selectEntityJSONFromHRef from "./getEntityJSONFromHRef"
+
 const isLinkWithStringHRef = isLink(isString)
-const selectEntityEmbeds = async <TLinks extends Links, TEmbeds extends string & keyof TLinks>(
-    client: ClientBase | undefined,
+
+const getEntityEmbeds = async <TLinks extends Links, TEmbeds extends string & keyof TLinks>(
+    client: S3Client,
     links: TLinks,
     embeds: readonly TEmbeds[],
     typeUserLabel: string,
@@ -15,7 +17,7 @@ const selectEntityEmbeds = async <TLinks extends Links, TEmbeds extends string &
         const propertyLink = links[property]
         const propertyJSON = `${JSON.stringify(property)}:`
         if (Array.isArray(propertyLink) && propertyLink.every(link => isLinkWithStringHRef(link))) {
-            return propertyJSON + (await selectEntitiesJSONFromLinks(client, propertyLink))
+            return propertyJSON + "[" + (await getEntitiesJSONFromLinks(client, propertyLink)).join(",") + "]"
         }
         if (isLinkWithStringHRef(propertyLink)) {
             return propertyJSON + (await selectEntityJSONFromHRef(client, propertyLink.href))
@@ -34,4 +36,5 @@ const selectEntityEmbeds = async <TLinks extends Links, TEmbeds extends string &
     })
     return `{${(await Promise.all(embedPromises)).join(",")}}`
 }
-export default selectEntityEmbeds
+
+export default getEntityEmbeds