+++ apps/api/src/operations/getCollection.ts
@@ -6,15 +6,19 @@ import { DataRequestHeaders } from "../headers/requests/DataRequestHeaders"
 import DATA_HEADERS from "../headers/responses/DATA_HEADERS"
 import checkAccept from "../mediaTypes/checkAccept"
 import { PgClientService } from "../services/PgClientService"
+import withPgClient from "../services/withPgClient"
 import validate from "../validation/validate"
 import { Operation } from "./Operation"
+
 const USER_MESSAGE = "There was a problem with an attempt to load a collection."
-export type GetCollectionParameters = DataRequestHeaders & Partial<CollectionParameters>
-export type GetCollectionService = PgClientService
+
+type GetCollectionParameters = DataRequestHeaders & Partial<CollectionParameters>
+
+type GetCollectionService = PgClientService
+
 const ensureExistence = async (service: PgClientService, uuid: string) => {
     if (uuid !== EMPTY_UUID) {
-        const client = await service.createPgClient()
-        try {
+        await withPgClient(service, async client => {
             const result = await client.query<{ uuid: UUID }>(
                 "SELECT uuid FROM collection WHERE uuid=$1::uuid LIMIT 1",
                 [uuid],
@@ -29,12 +33,11 @@ const ensureExistence = async (service: PgClientService, uuid: string) => {
                     },
                 ])
             }
-        } finally {
-            await service.deletePgClient(client)
-        }
+        })
     }
 }
-export const getCollection: Operation<GetCollectionParameters, GetCollectionService> = async (
+
+const getCollection: Operation<GetCollectionParameters, GetCollectionService> = async (
     { accept, ...queryAndPathParameters },
     service,
 ) => {
@@ -62,4 +65,5 @@ export const getCollection: Operation<GetCollectionParameters, GetCollectionServ
     // Unreachable.
     throw new Error()
 }
+
 export default getCollection