+++ apps/api/src/operations/getNodeLineage.ts
@@ -11,6 +11,7 @@ import {
 } from "@phylopic/api-models"
 import { normalizeUUID, UUID } from "@phylopic/utils"
 import { ClientBase } from "pg"
+import { S3Client } from "@aws-sdk/client-s3"
 import BUILD from "../build/BUILD"
 import checkBuild from "../build/checkBuild"
 import createBuildRedirect from "../build/createBuildRedirect"
@@ -18,17 +19,27 @@ import parseEntityJSONAndEmbed from "../entities/parseEntityJSONAndEmbed"
 import { DataRequestHeaders } from "../headers/requests/DataRequestHeaders"
 import checkAccept from "../mediaTypes/checkAccept"
 import checkListRedirect from "../pagination/checkListRedirect"
-import getListResult, { ListPageRow } from "../pagination/getListResult"
+import getPostgresListResult from "../pagination/getPostgresListResult"
+import { ListPageRow } from "../pagination/getListResult"
 import createPermanentRedirect from "../results/createPermanentRedirect"
 import { PgClientService } from "../services/PgClientService"
+import type { S3ClientService } from "../services/S3ClientService"
 import QueryConfigBuilder from "../sql/QueryConfigBuilder"
 import validate from "../validation/validate"
 import { Operation } from "./Operation"
-export type GetNodesParameters = DataRequestHeaders & NodeLineageParameters
-export type GetNodesService = PgClientService
+
+type GetNodesParameters = DataRequestHeaders & NodeLineageParameters
+
+type GetNodesService = PgClientService & S3ClientService
+
 const DEFAULT_TITLE = "[Unnamed]"
+
 const ITEMS_PER_PAGE = 48
+
 const USER_MESSAGE = "There was a problem with a request to list taxonomic groups."
+
+const VALID_EMBEDS = ["childNodes", "parentNode", "primaryImage"] as const
+
 const getQueryBuilder = (uuid: UUID, results: "total" | "href" | "json") => {
     const builder = new QueryConfigBuilder()
     const selection =
@@ -55,11 +66,13 @@ SELECT ${selection} FROM predecessors
     }
     return builder
 }
+
 const getTotalItems = (uuid: UUID) => async (client: ClientBase) => {
     const query = getQueryBuilder(uuid, "total").build()
     const queryResult = await client.query<{ total: string }>(query)
     return parseInt(queryResult.rows[0].total, 10) || 0
 }
+
 const getItemLinks =
     (uuid: UUID) =>
     async (client: ClientBase, offset: number, limit: number): Promise<readonly TitledLink[]> => {
@@ -71,6 +84,7 @@ const getItemLinks =
             title: title || DEFAULT_TITLE,
         }))
     }
+
 const fetchListPageRows =
     (uuid: UUID) =>
     async (client: ClientBase, offset: number, limit: number): Promise<readonly ListPageRow[]> => {
@@ -79,12 +93,13 @@ const fetchListPageRows =
         const queryResult = await client.query<{ json: string; title: string | null; uuid: UUID }>(queryBuilder.build())
         return queryResult.rows
     }
+
 const embedListPageRows =
-    (uuid: UUID) =>
+    () =>
     async (
         rows: readonly ListPageRow[],
         embeds: ReadonlyArray<string & keyof NodeEmbedded>,
-        client?: ClientBase,
+        client: S3Client,
     ): Promise<readonly Readonly<[TitledLink, string]>[]> => {
         if (!embeds.length) {
             return rows.map(({ json, title, uuid }) => [
@@ -101,7 +116,8 @@ const embedListPageRows =
             }),
         )
     }
-export const getNodeLineage: Operation<GetNodesParameters, GetNodesService> = async (
+
+const getNodeLineage: Operation<GetNodesParameters, GetNodesService> = async (
     { accept, ...queryAndPathParameters },
     service,
 ) => {
@@ -117,18 +133,19 @@ export const getNodeLineage: Operation<GetNodesParameters, GetNodesService> = as
         return createPermanentRedirect(path, { ...queryParameters, uuid: normalizedUUID })
     }
     checkBuild(queryParameters.build, USER_MESSAGE)
-    return await getListResult({
-        embedListPageRows: embedListPageRows(normalizedUUID),
-        fetchListPageRows: fetchListPageRows(normalizedUUID),
-        getItemLinks: getItemLinks(normalizedUUID),
-        getTotalItems: getTotalItems(normalizedUUID),
+    return await getPostgresListResult({
         itemsPerPage: ITEMS_PER_PAGE,
         listPath: path,
-        service,
         listQuery: queryParameters,
         page: queryParameters.page,
         userMessage: USER_MESSAGE,
-        validEmbeds: ["childNodes", "parentNode", "primaryImage"],
+        validEmbeds: VALID_EMBEDS,
+        embedListPageRows: embedListPageRows(),
+        fetchListPageRows: fetchListPageRows(normalizedUUID),
+        getItemLinks: getItemLinks(normalizedUUID),
+        getTotalItems: getTotalItems(normalizedUUID),
+        service,
     })
 }
+
 export default getNodeLineage