+++ apps/api/src/search/normalizeSearchQuery.ts
@@ -1,7 +1,7 @@
 import { normalizeQuery } from "@phylopic/api-models"
 import APIError from "../errors/APIError"
 import MIN_QUERY_LENGTH from "./MIN_QUERY_LENGTH"
-export const normalizeSearchQuery = (query: string) => {
+const normalizeSearchQuery = (query: string) => {
     query = normalizeQuery(query)
     if (query.length < MIN_QUERY_LENGTH) {
         throw new APIError(400, [