+++ packages/s3-entities/src/ENTITIES_BUCKET.ts
@@ -1,0 +1,1 @@
+export const ENTITIES_BUCKET = process.env.ENTITIES_BUCKET ?? "entities.phylopic.org"